+++ CHANGES.rst
@@ -17,7 +17,7 @@ Internals
 * Definitions for symbols ``CurrentContext`` and ``ContextPath[]`` are mirrored in the ``mathics.core.definitions.Definitions`` object for faster access. 
 * ``FullForm[List[...]]`` now is shown as ``{...}`` according to the WL standard.
 * ``Expression.is_numeric()`` accepts an ``Evaluation`` object as a parameter;  the definitions attribute of that is used.
-* To numerify expressions, the function ``apply_N`` was introduced in module ``mathics.builtin.numeric`` to speed up the critical built-in function
+* To numerify expressions, the function ``apply_N`` was introduced in module ``mathics.core.evaluators`` to speed up the critical built-in function
 ``N``. Its use instead of the idiom ``Expression("N", expr, prec).evaluate(evaluation)`` makes the evaluation faster.
 * A bug was fixed relating to the order in which ``mathics.core.definitions`` stores the rules.
 * A bug comming from a failure in the order in which `mathics.core.definitions` stores the rules was fixed.
@@ -33,6 +33,9 @@ Internals
 * ``Cylinder`` rendering is implemented in Asymptote.
 * `N[_,_,Method->method]` was reworked (Issue #137).
 * ``N.apply_with_prec`` and ``apply_N`` was reworked to reduce the number of calls to the ``evaluate`` method.
+* `numerify` is now a non-member function and uses `apply_N`.
+
+
   
 Compatibility
 +++++++++++++